Add files via upload
A comprehensive survey was conducted to investigate the key factors influencing the investment ideologies of the youth. Through this survey,  primary data was collected from a diverse group of respondents. Following data collection, rigorous statistical analysis was performed to extract meaningful patterns and correlations within the dataset. The findings were then visually presented through informative graphs and charts, providing clear and insightful conclusions that shed light on the dynamics shaping the investment choices of young individuals. This research not only expanded our knowledge of the subject but also offered valuable insights that can inform investment strategies and financial education programs tailored to the needs of today's youth.